Creates/Adds/Modifies - Adds Educator Class - Adds Human Resources Module - Refactors student, teacher, and apprentice to make more dry
