+++ human_resources.rb
@@ -1,0 +1,28 @@
+module HumanResources
+
+  def offer_high_five
+    "High five!"
+  end
+
+  def set_performance_rating(rating)
+    response = ""
+    if rating > 90
+      receive_raise(@target_raise)
+      response = "Yay, I'm a great employee!"
+    else
+      response += "Oh, well -- thanks to this actionable, specific, and kind "
+      response += "feedback, I'll do better next time."
+    end
+    response
+  end
+
+  def salary=(new_salary)
+    puts "This better be good!"
+    @salary = new_salary
+  end
+
+  def receive_raise(raise)
+    @salary += raise
+  end
+
+end
