+++ spec/student_spec.rb
@@ -1,4 +1,5 @@
 require_relative '../student'
+require_relative '../human_resources'
 
 describe Student do
   let(:amir) { Student.new({name: "Amir Fayek", age: 21}) }
