+++ student.rb
@@ -1,16 +1,18 @@
 class Student
-  attr_reader :age, :phase
-  attr_accessor :name
+  # attr_reader :age, :phase
+  # attr_accessor :name
+
+  include HumanResources
 
   def initialize(options = {})
     @phase = 1
-    @age = options.fetch(:age, 0)
-    @name = options.fetch(:name, "")
+    # @age = options.fetch(:age, 0)
+    # @name = options.fetch(:name, "")
   end
 
-  def offer_high_five
-    "High five!"
-  end
+  # def offer_high_five
+  #   "High five!"
+  # end
 
   def set_phase(num)
     response = ""
